Update README.md
The projects that I want to make are the apps that I usually use. One of which would be a simple HRV-Resonant Breathing app. I hacked about in PowerPoint to give myself an idea of what I'd need to code, but it wasn't working. After some time, I decided to search. Lo and Behold, the incomparable Brad Traversy had posted a tutorial. 

This was a great learning ground. It was the app I wanted to build, but not quite. Brad's had quicker-paced breathing, and it included a breath-hold. I wanted an app that trained you to breath at 5 breaths per minute at an inhale:exhale ratio of 4:6. The challenges here were learning how to modify the javascript and the gradient in order to get everything to match up. 

I also wanted a simple sans-serif (Josefin Sans), and a personalised background image. The image is a photograph taken by my wife.  The image is of freshly fallen gingko leaves.

Ways to take this further would be to insert the image into a colour picker and to come up with a better colour palette for the gradient. I may in future do so. 

The headline isn't meant to be rude, but rather is an homage to that Baron of Breathwork, Wim Hof.